chore: move authenticators into native
